% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDCRNATools-package.R
\docType{data}
\name{DEGAll}
\alias{DEGAll}
\title{Output of \code{\link{gdcDEAnalysis}} for downstream analysis}
\description{
Output of \code{\link{gdcDEAnalysis}} for downstream analysis
}
\keyword{datasets}
