% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDCRNATools-package.R
\docType{data}
\name{enrichOutput}
\alias{enrichOutput}
\title{Output of \code{\link{gdcEnrichAnalysis}} for visualization}
\description{
Output of \code{\link{gdcEnrichAnalysis}} for visualization
}
\keyword{datasets}
