% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdcDEGVisulization.R
\name{gdcBarPlot}
\alias{gdcBarPlot}
\title{Bar plot of differentially expressed genes/miRNAs}
\usage{
gdcBarPlot(deg, angle = 0, data.type)
}
\arguments{
\item{deg}{a dataframe generated from \code{\link{gdcDEReport}} containing DE genes/miRNAs ids, logFC, etc.}

\item{angle}{a numeric value specifying the angle of text on x-axis. Default is \code{0}}

\item{data.type}{one of \code{'RNAseq'} and \code{'miRNAs'}}
}
\value{
A bar plot
}
\description{
A bar plot showing the number of down-regulated and up-regulated DE genes/miRNAs of different biotypes
}
\author{
Ruidong Li and Han Qu
}
