% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdcCEAnalysis.R
\name{gdcCEAnalysis}
\alias{gdcCEAnalysis}
\title{Competing endogenous RNAs (ceRNAs) analysis}
\usage{
gdcCEAnalysis(lnc, pc, deMIR = NULL, lnc.targets = "starBase",
  pc.targets = "starBase", rna.expr, mir.expr)
}
\arguments{
\item{lnc}{a vector of Ensembl long non-coding gene ids}

\item{pc}{a vector of Ensembl protein coding gene ids}

\item{deMIR}{a vector of differentially expressed miRNAs. Default is \code{NULL}}

\item{lnc.targets}{a character string specifying the database of miRNA-lncRNA interactions.
Should be one of \code{'spongeScan'}, \code{'starBase'}, and \code{'miRcode'}. Default is \code{'starBase'}. \cr\cr
Or a \code{list} of miRNA-lncRNA interactions generated by users}

\item{pc.targets}{a character string specifying the database of miRNA-lncRNA interactions.
Should be one of \code{'spongeScan'}, \code{'starBase'}, and \code{'miRcode'}. Default is \code{'starBase'}. \cr\cr
Or a \code{list} of miRNA-lncRNA interactions generated by users}

\item{rna.expr}{\code{\link[limma]{voom}} transformed gene expression data}

\item{mir.expr}{\code{\link[limma]{voom}} transformed mature miRNA expression data}
}
\value{
A dataframe containing ceRNA pairs, expression correlation between lncRNA and mRNA,
  the number and hypergeometric significance of shared miRNAs, regulation similarity score, 
  and the mean sensitity correlation (the difference between Pearson correlation and partial correlation)
  of multiple lncRNA-miRNA-mRNA triplets, etc.
}
\description{
Identify ceRNAs by (1) number of shared miRNAs between lncRNA and mRNA; 
  (2) expression correlation of lncRNA and mRNA;
  (3) regulation similarity of shared miRNAs on lncRNA and mRNA;
  (4) sensitivity correlation
}
\references{
Paci P, Colombo T, Farina L. Computational analysis identifies a sponge interaction network between long non-coding RNAs and messenger RNAs in human breast cancer. BMC systems biology. 2014 Jul 17;8(1):83.
}
\author{
Ruidong Li and Han Qu
}
