% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdcCorPlot.R
\name{gdcCorPlot}
\alias{gdcCorPlot}
\title{Correlation plot of two genes/miRNAs}
\usage{
gdcCorPlot(gene1, gene2, rna.expr, metadata)
}
\arguments{
\item{gene1}{an Ensembl gene id or miRBase v21 mature miRNA id}

\item{gene2}{an Ensembl gene id or miRBase v21 mature miRNA id}

\item{rna.expr}{\code{\link[limma]{voom}} transformed expression data}

\item{metadata}{metadata parsed from \code{\link{gdcParseMetadata}}}
}
\value{
A scatter plot with line of best fit
}
\description{
Scatter plot showing the expression correlation between two genes/miRNAs
}
\author{
Ruidong Li and Han Qu
}
