% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdcDEAnalysis.R
\name{gdcDEAnalysis}
\alias{gdcDEAnalysis}
\title{Differential gene expression analysis}
\usage{
gdcDEAnalysis(counts, group, comparison, method = "limma", n.cores = NULL,
  filter = TRUE)
}
\arguments{
\item{counts}{a dataframe or numeric matrix of raw counts data generated from \code{\link{gdcRNAMerge}}}

\item{group}{a vector giving the group that each sample belongs to}

\item{comparison}{a character string specifying the two groups being compared. \cr
Example: \code{comparison='PrimaryTumor-SolidTissueNormal'}}

\item{method}{one of \code{'limma'}, \code{'edgeR'}, and \code{'DESeq2'}. Default is \code{'limma'} \cr
Note: It may takes long time for \code{method='DESeq2'} with a single core}

\item{n.cores}{a numeric value of cores to be used for \code{method='DESeq2'} to accelate the analysis process. Default is \code{NULL}}

\item{filter}{logical, whether to filter out low expression genes. If \code{TRUE}, only genes 
with \code{cpm > 1} in more than half of the samples will be kept. Default is \code{TRUE}}
}
\value{
A dataframe containing Ensembl gene ids/miRBase v21 mature miRNA ids, gene symbols, biotypes, fold change on the log2 scale,
  p value, and FDR etc. of all genes/miRNAs of analysis.
}
\description{
Performs differential gene expression analysis by \pkg{limma}, \pkg{edgeR}, and \pkg{DESeq2}
}
\note{
It may takes long time for \code{method='DESeq2'} with a single core. Please use multiple cores if possible
}
\references{
Robinson MD, McCarthy DJ, Smyth GK. edgeR: a Bioconductor package for differential expression analysis of digital gene expression data. Bioinformatics. 2010 Jan 1;26(1):139-40. \cr
  Ritchie ME, Phipson B, Wu D, Hu Y, Law CW, Shi W, Smyth GK. limma powers differential expression analyses for RNA-sequencing and microarray studies. Nucleic acids research. 2015 Jan 20;43(7):e47-e47. \cr
  Love MI, Huber W, Anders S. Moderated estimation of fold change and dispersion for RNA-seq data with DESeq2. Genome biology. 2014 Dec 5;15(12):550.
}
\author{
Ruidong Li and Han Qu
}
