% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdcDEAnalysis.R
\name{gdcDEReport}
\alias{gdcDEReport}
\title{Report differentially expressed genes/miRNAs}
\usage{
gdcDEReport(deg, gene.type = "all", fc = 2, pval = 0.01)
}
\arguments{
\item{deg}{A dataframe of DE analysis result from \code{\link{gdcDEAnalysis}}}

\item{gene.type}{one of \code{'all'}, \code{'long_non_coding'}, \code{'protein_coding'}, and \code{'miRNAs'}. Default is \code{'all'}}

\item{fc}{a numeric value specifying the threshold of fold change}

\item{pval}{a nuemric value specifying the threshold of p value}
}
\value{
A dataframe or numeric matrix of differentially expressed genes/miRNAs
}
\description{
Report genes/miRNAs that are differentially expressed satisfying a given threshold
}
\author{
Ruidong Li and Han Qu
}
