% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdcExportNetwork.R
\name{gdcExportNetwork}
\alias{gdcExportNetwork}
\title{Export network for Cytoscape}
\usage{
gdcExportNetwork(ceNetwork, net)
}
\arguments{
\item{ceNetwork}{a dataframe generated from \code{\link{gdcCEAnalysis}}}

\item{net}{one of \code{'nodes'} and \code{'edges'}}
}
\value{
A dataframe of nodes or edges
}
\description{
Export nodes and edges of ce network for \strong{Cytoscape} visualization
}
\author{
Ruidong Li and Han Qu
}
