% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdcFilterSamples.R
\name{gdcFilterDuplicate}
\alias{gdcFilterDuplicate}
\title{Filter out duplicated samples}
\usage{
gdcFilterDuplicate(metadata)
}
\arguments{
\item{metadata}{metadata parsed from \code{\link{gdcParseMetadata}}}
}
\value{
A filtered dataframe of metadata without duplicated samples
}
\description{
Filter out samples that are sequenced for two or more times
}
\examples{
####### Parse metadata by project id and data type #######
metaMatrix <- gdcParseMetadata(project.id='TARGET-RT', data.type='RNAseq')
metaMatrix <- gdcFilterDuplicate(metadata=metaMatrix)
}
\author{
Ruidong Li and Han Qu
}
