% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdcDEGVisulization.R
\name{gdcHeatmap}
\alias{gdcHeatmap}
\title{Heatmap of differentially expressed genes/miRNAs}
\usage{
gdcHeatmap(deg.id, metadata, rna.expr)
}
\arguments{
\item{deg.id}{a vector of Ensembl gene ids or miRBase v21 mature miRNA ids}

\item{metadata}{metadata parsed from \code{\link{gdcParseMetadata}}}

\item{rna.expr}{\code{\link[limma]{voom}} transformed expression data}
}
\value{
A heatmap with rows are DE genes/miRNAs and columns are samples.
  \emph{Solid Tissue Normal} samples are labeled with blue and \emph{Primary Tumor} samples are labeled with red
}
\description{
A heatmap showing unsupervised hierarchical clustering of 
  DE genes/miRNAs by \code{\link[gplots]{heatmap.2}} in the \pkg{gplots} package
}
\author{
Ruidong Li and Han Qu
}
