% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdcKMPlot.R
\name{gdcKMPlot}
\alias{gdcKMPlot}
\title{Kaplan Meier plot}
\usage{
gdcKMPlot(gene, rna.expr, metadata, sep = "median")
}
\arguments{
\item{gene}{an Ensembl gene id}

\item{rna.expr}{\code{\link[limma]{voom}} transformed expression data}

\item{metadata}{metadata parsed from \code{\link{gdcParseMetadata}}}

\item{sep}{a character string specifying which point should be used to separate low-expression and high-expression groups.
Possible values are \code{'1stQu'}, \code{'mean'}, \code{'median'}, and \code{'3rdQu'}. Default is \code{'median'}}
}
\value{
A plot of Kaplan Meier survival curve
}
\description{
Plot Kaplan Meier survival curve
}
\author{
Ruidong Li and Han Qu
}
