% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdcMatchSamples.R
\name{gdcMatchSamples}
\alias{gdcMatchSamples}
\title{Match samples in metadata and expression matrix}
\usage{
gdcMatchSamples(metadata, rna.expr)
}
\arguments{
\item{metadata}{metadata parsed from \code{\link{gdcParseMetadata}}}

\item{rna.expr}{\code{\link[limma]{voom}} transformed expression data}
}
\value{
A logical value. If \code{TRUE}, all the samples matched
}
\description{
Check if samples in the metadata and expression data match
}
\author{
Ruidong Li and Han Qu
}
