% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdcRNADownload.R
\name{gdcRNADownload}
\alias{gdcRNADownload}
\title{Download RNA data in GDC}
\usage{
gdcRNADownload(manifest = NULL, project.id, data.type, directory = "Data",
  write.manifest = FALSE)
}
\arguments{
\item{manifest}{menifest file that is downloaded from the GDC cart. If provided, files whose UUIDs are in the manifest file 
will be downloaded via gdc-client, otherwise, \code{project} and \code{data.type} arguments should be provided to download 
data by the API method implemented in the \pkg{TCGAbiolinks} package. Default is \code{NULL}}

\item{project.id}{project id in GDC}

\item{data.type}{one of \code{'RNAseq'} and \code{'miRNAs'}}

\item{directory}{the folder to save downloaded files. Default is \code{'Data'}}

\item{write.manifest}{logical, whether to write out the manifest file}
}
\description{
Download gene expression quantification and isoform expression quantification data from GDC 
  either by providing the manifest file or by the API method developed in \pkg{TCGAbiolinks} package
}
\author{
Ruidong Li and Han Qu
}
