% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdcRNAMerge.R
\name{gdcRNAMerge}
\alias{gdcRNAMerge}
\title{Merge RNA/miRNAs raw counts data}
\usage{
gdcRNAMerge(metadata, path, data.type)
}
\arguments{
\item{metadata}{metadata parsed from \code{\link{gdcParseMetadata}}}

\item{path}{path to downloaded files for merging}

\item{data.type}{one of \code{'RNAseq'} and \code{'miRNAs'}}
}
\value{
A dataframe or numeric matrix of raw counts data with rows are genes or miRNAs and columns are samples
}
\description{
Merge raw counts data that is downloaded from GDC to a single expression matrix
}
\author{
Ruidong Li and Han Qu
}
