% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDCRNATools-package.R
\docType{package}
\name{GDCRNATools-package}
\alias{GDCRNATools-package}
\alias{GDCRNATools}
\title{This is an easy-to-use package for downloading, organizing, 
and integrative analyzing RNA expression data in GDC with an 
emphasis on deciphering the lncRNA-mRNA related ceRNA regulatory 
network in cancer.}
\description{
This is an easy-to-use package for downloading, organizing, 
and integrative analyzing RNA expression data in GDC with an 
emphasis on deciphering the lncRNA-mRNA related ceRNA regulatory 
network in cancer.
}
