% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdcSurvivalAnalysis.R
\name{gdcSurvivalAnalysis}
\alias{gdcSurvivalAnalysis}
\title{Univariate survival analysis of multiple genes}
\usage{
gdcSurvivalAnalysis(gene, rna.expr, metadata, method = "coxph",
  sep = "median")
}
\arguments{
\item{gene}{a vector of Ensembl gene ids}

\item{rna.expr}{\code{\link[limma]{voom}} transformed expression data}

\item{metadata}{metadata parsed from \code{\link{gdcParseMetadata}}}

\item{method}{method for survival analysis. Possible values are \code{'coxph'} and \code{'KM'}. Default is \code{'coxph'}}

\item{sep}{which point should be used to separate low-expression and high-expression groups for \code{method='KM'}.
Possible values are \code{'1stQu'}, \code{'mean'}, \code{'median'}, and \code{'3rdQu'}. Default is \code{'median'}}
}
\value{
A dataframe or numeric matrix of hazard ratio, 95\% confidence interval, p value, and FDR
}
\description{
Univariate Cox Proportional-Hazards and Kaplan Meier survival analysis of a vector of genes
}
\references{
Therneau TM, Lumley T. Package ‘survival’. \cr
  Andersen PK, Gill RD. Cox's regression model for counting processes: a large sample study. The annals of statistics. 1982 Dec 1:1100-20. \cr
  Therneau TM, Grambsch PM. Extending the Cox model. Edited by P. Bickel, P. Diggle, S. Fienberg, K. Krickeberg. 2000:51. \cr
  Harrington DP, Fleming TR. A class of rank test procedures for censored survival data. Biometrika. 1982 Dec 1;69(3):553-66.
}
\author{
Ruidong Li and Han Qu
}
