% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdcDEGVisulization.R
\name{gdcVolcanoPlot}
\alias{gdcVolcanoPlot}
\title{Volcano plot of differentially expressed genes/miRNAs}
\usage{
gdcVolcanoPlot(deg.all, fc = 2, pval = 0.01)
}
\arguments{
\item{deg.all}{a dataframe generated from \code{\link{gdcDEAnalysis}} containing all genes of analysis no matter they are differentially expressed or not}

\item{fc}{a numeric value specifying the threshold of fold change}

\item{pval}{a nuemric value specifying the threshold of p value}
}
\value{
A volcano plot
}
\description{
A volcano plot showing differentially expressed genes/miRNAs
}
\author{
Ruidong Li and Han Qu
}
