% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdcShinyPlot.R
\name{shinyKMPlot}
\alias{shinyKMPlot}
\title{Shiny Kaplan Meier (KM) plot}
\usage{
shinyKMPlot(gene, rna.expr, metadata)
}
\arguments{
\item{gene}{a vector of Ensembl gene ids}

\item{rna.expr}{\code{\link[limma]{voom}} transformed expression data}

\item{metadata}{metadata parsed from \code{\link{gdcParseMetadata}}}
}
\value{
a local webpage for visualization of KM plots
}
\description{
A simple \pkg{shiny} app to show KM survival curves on local web browser
}
\author{
Ruidong Li and Han Qu
}
