% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdcClinical.R
\name{gdcClinicalDownload}
\alias{gdcClinicalDownload}
\title{Download clinical data in GDC}
\usage{
gdcClinicalDownload(manifest = NULL, project.id, directory = "Clinical",
  write.manifest = FALSE)
}
\arguments{
\item{manifest}{menifest file that is downloaded from the GDC cart. If provided, files whose UUIDs are in the manifest file 
will be downloaded via gdc-client, otherwise, \code{project} argument should be provided to download 
data by the API method implemented in the \pkg{TCGAbiolinks} package. Default is \code{NULL}}

\item{project.id}{project id in GDC}

\item{directory}{the folder to save downloaded files. Default is \code{'Clinical'}}

\item{write.manifest}{logical, whether to write out the manifest file}
}
\description{
Download clinical data in GDC 
  either by providing the manifest file or by the API method developed in \pkg{TCGAbiolinks} package
}
\author{
Ruidong Li and Han Qu
}
